+++ pkg/utils/kubernetes/resources/services.go
@@ -61,10 +61,11 @@ func GenerateNewIngressServiceForDataPlane(dataplane *operatorv1beta1.DataPlane,
 			Selector: map[string]string{
 				"app": dataplane.Name,
 			},
-			Ports:                 DefaultDataPlaneIngressServicePorts,
-			ExternalTrafficPolicy: getDataPlaneIngressServiceExternalTrafficPolicy(dataplane),
+			Ports: DefaultDataPlaneIngressServicePorts,
 		},
 	}
+
+	setDataPlaneIngressServiceExternalTrafficPolicy(dataplane, svc)
 	LabelObjectAsDataPlaneManaged(svc)
 
 	for _, opt := range opts {
@@ -112,12 +113,18 @@ func getDataPlaneIngressServiceType(dataplane *operatorv1beta1.DataPlane) corev1
 	return dataplane.Spec.Network.Services.Ingress.Type
 }
 
-func getDataPlaneIngressServiceExternalTrafficPolicy(dataplane *operatorv1beta1.DataPlane) corev1.ServiceExternalTrafficPolicy {
-	if dataplane == nil || dataplane.Spec.Network.Services == nil {
-		return corev1.ServiceExternalTrafficPolicyCluster
+func setDataPlaneIngressServiceExternalTrafficPolicy(
+	dataplane *operatorv1beta1.DataPlane,
+	svc *corev1.Service,
+) {
+	if dataplane == nil ||
+		dataplane.Spec.Network.Services == nil ||
+		dataplane.Spec.Network.Services.Ingress == nil ||
+		dataplane.Spec.Network.Services.Ingress.ExternalTrafficPolicy == "" {
+		return
 	}
 
-	return dataplane.Spec.Network.Services.Ingress.ExternalTrafficPolicy
+	svc.Spec.ExternalTrafficPolicy = dataplane.Spec.Network.Services.Ingress.ExternalTrafficPolicy
 }
 
 // ServiceOpt is an option function for a Service.